Fix #205 (Uninstall fix)
- Moved some Command-Handling methods to their correct location
